added git log and mv
