+++ completers/git_completer/cmd/mv.go
@@ -1,20 +1,24 @@
 package cmd
 
 import (
+	"github.com/rsteube/carapace"
 	"github.com/spf13/cobra"
 )
 
 var mvCmd = &cobra.Command{
 	Use:   "mv",
 	Short: "Move or rename a file, a directory, or a symlink",
-	Run: func(cmd *cobra.Command, args []string) {
-	},
+	Run:   func(cmd *cobra.Command, args []string) {},
 }
 
 func init() {
-	mvCmd.Flags().BoolP("force", "f", false, "force move/rename even if target exists")
+	carapace.Gen(mvCmd).Standalone()
+
 	mvCmd.Flags().BoolS("k", "k", false, "skip move/rename errors")
 	mvCmd.Flags().BoolP("dry-run", "n", false, "dry run")
+	mvCmd.Flags().BoolP("force", "f", false, "force move/rename even if target exists")
 	mvCmd.Flags().BoolP("verbose", "v", false, "be verbose")
 	rootCmd.AddCommand(mvCmd)
+
+	carapace.Gen(mvCmd).PositionalAnyCompletion(carapace.ActionFiles(""))
 }
